align exports & module boundaries for code splitting
