+++ ui/keyboardMove/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "description": "lichess.org keyboard move",
   "type": "module",
-  "module": "dist/ctrl.js",
-  "types": "dist/ctrl.d.ts",
+  "module": "dist/exports.js",
+  "types": "dist/exports.d.ts",
   "keywords": [
     "chess",
     "lichess",
